[core] move the LockedTXN class out of txpool so it may be reused
